Organize code to conform to new RTOS architecture
